+++ tests/test_e2e/test_run_all_ic.py
@@ -150,7 +150,7 @@ def test_get_capability_from_app_config_and_run_ic(config_fixture_name, request,
     else:
         raise ValueError("Test should be rewritten if more than two datasets used.")
 
-    if capability.supports_metrics == Number.ONE:
+    if capability.supports_metrics in {Number.ONE, Number.MANY}:
         metrics = [metric_ic]
     elif capability.supports_metrics == Number.ZERO:
         metrics = []